requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.0.2
Flask-Cors==4.0.0
fluidsynth==0.2
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.4
MIDIUtil==1.2.1
multidict==6.1.0
numpy==2.0.2
ollama==0.1.6
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.42.0
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.0.1
pygame==2.6.1
pyparsing==3.2.1
pyppeteer==2.0.0
pyshark==0.6
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
scipy==1.13.1
sniffio==1.3.1
soupsieve==2.7
termcolor==2.5.0
text-unidecode==1.3
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
websockets==10.4
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
